// api/models/AbilitiesRequiredItems.ts
/* tslint:disable */
/* eslint-disable */
/**
 * standard public schema
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 13.0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AbilitiesRequiredItems
 */
export interface AbilitiesRequiredItems {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     * This is a Foreign Key to `abilities.id`.<fk table='abilities' column='id'/>
     * @type {number}
     * @memberof AbilitiesRequiredItems
     */
    abilityId: number;
    /**
     * Note:
     * This is a Primary Key.<pk/>
     * @type {string}
     * @memberof AbilitiesRequiredItems
     */
    itemTree: string;
    /**
     * 
     * @type {number}
     * @memberof AbilitiesRequiredItems
     */
    requiredQuantity?: number;
}

/**
 * Check if a given object implements the AbilitiesRequiredItems interface.
 */
export function instanceOfAbilitiesRequiredItems(value: object): value is AbilitiesRequiredItems {
    if (!('abilityId' in value) || value['abilityId'] === undefined) return false;
    if (!('itemTree' in value) || value['itemTree'] === undefined) return false;
    return true;
}

export function AbilitiesRequiredItemsFromJSON(json: any): AbilitiesRequiredItems {
    return AbilitiesRequiredItemsFromJSONTyped(json, false);
}

export function AbilitiesRequiredItemsFromJSONTyped(json: any, ignoreDiscriminator: boolean): AbilitiesRequiredItems {
    if (json == null) {
        return json;
    }
    return {
        
        'abilityId': json['ability_id'],
        'itemTree': json['item_tree'],
        'requiredQuantity': json['required_quantity'] == null ? undefined : json['required_quantity'],
    };
}

export function AbilitiesRequiredItemsToJSON(json: any): AbilitiesRequiredItems {
    return AbilitiesRequiredItemsToJSONTyped(json, false);
}

export function AbilitiesRequiredItemsToJSONTyped(value?: AbilitiesRequiredItems | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'ability_id': value['abilityId'],
        'item_tree': value['itemTree'],
        'required_quantity': value['requiredQuantity'],
    };
}