// api/models/AbilitiesRequiredLevels.ts
/* tslint:disable */
/* eslint-disable */
/**
 * standard public schema
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 13.0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface AbilitiesRequiredLevels
 */
export interface AbilitiesRequiredLevels {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     * This is a Foreign Key to `abilities.id`.<fk table='abilities' column='id'/>
     * @type {number}
     * @memberof AbilitiesRequiredLevels
     */
    abilityId: number;
    /**
     * Note:
     * This is a Primary Key.<pk/>
     * This is a Foreign Key to `skills.id`.<fk table='skills' column='id'/>
     * @type {number}
     * @memberof AbilitiesRequiredLevels
     */
    skillId: number;
    /**
     * 
     * @type {number}
     * @memberof AbilitiesRequiredLevels
     */
    requiredLevel?: number;
}

/**
 * Check if a given object implements the AbilitiesRequiredLevels interface.
 */
export function instanceOfAbilitiesRequiredLevels(value: object): value is AbilitiesRequiredLevels {
    if (!('abilityId' in value) || value['abilityId'] === undefined) return false;
    if (!('skillId' in value) || value['skillId'] === undefined) return false;
    return true;
}

export function AbilitiesRequiredLevelsFromJSON(json: any): AbilitiesRequiredLevels {
    return AbilitiesRequiredLevelsFromJSONTyped(json, false);
}

export function AbilitiesRequiredLevelsFromJSONTyped(json: any, ignoreDiscriminator: boolean): AbilitiesRequiredLevels {
    if (json == null) {
        return json;
    }
    return {
        
        'abilityId': json['ability_id'],
        'skillId': json['skill_id'],
        'requiredLevel': json['required_level'] == null ? undefined : json['required_level'],
    };
}

export function AbilitiesRequiredLevelsToJSON(json: any): AbilitiesRequiredLevels {
    return AbilitiesRequiredLevelsToJSONTyped(json, false);
}

export function AbilitiesRequiredLevelsToJSONTyped(value?: AbilitiesRequiredLevels | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'ability_id': value['abilityId'],
        'skill_id': value['skillId'],
        'required_level': value['requiredLevel'],
    };
}