// api/models/Backgrounds.ts
export interface Backgrounds {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     * @type {number}
     * @memberof Backgrounds
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Backgrounds
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Backgrounds
     */
    description?: string;
    /**
     * 
     * @type {number}
     * @memberof Backgrounds
     */
    bonusSpeed?: number;
    /**
     * 
     * @type {number}
     * @memberof Backgrounds
     */
    bonusHealth?: number;
    /**
     * 
     * @type {number}
     * @memberof Backgrounds
     */
    bonusMana?: number;
    /**
     * 
     * @type {number}
     * @memberof Backgrounds
     */
    bonusStrength?: number;
    /**
     * 
     * @type {number}
     * @memberof Backgrounds
     */
    bonusDexterity?: number;
    /**
     * 
     * @type {number}
     * @memberof Backgrounds
     */
    bonusIntelligence?: number;
    /**
     * 
     * @type {number}
     * @memberof Backgrounds
     */
    bonusCharisma?: number;
    /**
     * 
     * @type {number}
     * @memberof Backgrounds
     */
    bonusWisdom?: number;
    /**
     * 
     * @type {number}
     * @memberof Backgrounds
     */
    bonusConstitution?: number;
    /**
     * 
     * @type {string}
     * @memberof Backgrounds
     */
    image?: string;
}