// api/models/Skills.ts
/* tslint:disable */
/* eslint-disable */
/**
 * standard public schema
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 13.0.4
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Skills
 */
export interface Skills {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     * @type {number}
     * @memberof Skills
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Skills
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Skills
     */
    description?: string;
    /**
     * 
     * @type {string}
     * @memberof Skills
     */
    type?: string;
    /**
     * 
     * @type {string}
     * @memberof Skills
     */
    image?: string;
}

/**
 * Check if a given object implements the Skills interface.
 */
export function instanceOfSkills(value: object): value is Skills {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    return true;
}

export function SkillsFromJSON(json: any): Skills {
    return SkillsFromJSONTyped(json, false);
}

export function SkillsFromJSONTyped(json: any, ignoreDiscriminator: boolean): Skills {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'description': json['description'] == null ? undefined : json['description'],
        'type': json['type'] == null ? undefined : json['type'],
        'image': json['image'] == null ? undefined : json['image'],
    };
}

export function SkillsToJSON(json: any): Skills {
    return SkillsToJSONTyped(json, false);
}

export function SkillsToJSONTyped(value?: Skills | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'description': value['description'],
        'type': value['type'],
        'image': value['image'],
    };
}